factor out cli to its own module
